move files + features reduction
