+++ setup.cfg
@@ -33,6 +33,7 @@ python_requires = >=3.8
 scripts =
   src/Caribou_pipeline.py
   src/Caribou_kmers.py
+  src/Caribou_reduce_features.py
   src/Caribou_extraction.py
   src/Caribou_classification.py
   src/Caribou_extraction_train_cv.py
@@ -45,11 +46,6 @@ scripts =
   data/subset_merge.py
 
 [options.package_data]
-KMC =
-  src/data/KMC/bin/*
-  src/data/KMC/include/kmc_runner.h
-faSplit =
-  src/data/faSplit
 KronaTools =
   src/outputs/KronaTools/img/*
   src/outputs/KronaTools/lib/*
