--- src/__init__.py
@@ -1,18 +1,0 @@
-###########################################
-##                                       ##
-##      2021(C) Nicolas de Montigny      ##
-##                                       ##
-##      Laboratoire Bioinformatique      ##
-##      Pr. Abdoulaye Diallo Group       ##
-##                  &                    ##
-## Laboratoire de microbiome des plantes ##
-##       Pr. Steven Kembel Group         ##
-##                UQAM                   ##
-##                                       ##
-###########################################
-
-__author__ = "Nicolas de Montigny"
-__date__   = "20210601"
-__version__ = "1.0.0"
-
-__all__ = ['data','environments','models','outputs','Caribou','utils','supplement','K-mers_extract_dataset','Bacteria_classification_train_cv','Bacteria_extraction_train_cv']
