--- src/data/KMC/bin/__init__.py
@@ -1,16 +1,0 @@
-###########################################
-##                                       ##
-##      2021(C) Nicolas de Montigny      ##
-##                                       ##
-##      Laboratoire Bioinformatique      ##
-##      Pr. Abdoulaye Diallo Group       ##
-##                  &                    ##
-## Laboratoire de microbiome des plantes ##
-##       Pr. Steven Kembel Group         ##
-##                UQAM                   ##
-##                                       ##
-###########################################
-
-__author__ = "Nicolas de Montigny"
-
-__all__ = ['kmc','kmc_dump','kmc_tools','libkmc_core.a']
